requirements.txt
django-allauth

Django


gunicorn

django-bootstrap-v5

django-heroku

python-dotenv